+++ examples/MultiArrayTester/Constants.h
@@ -8,7 +8,7 @@ enum {X_SIZE=3};
 enum {Y_SIZE=4};
 enum {Z_SIZE=5};
 
-extern const unsigned int baudrate;
+extern const long baudrate;
 
 extern const int x_default[X_SIZE];
 extern const int y_default[Y_SIZE];
